refactor: precondition fixes
