+++ src/angular/autocomplete/autocomplete.spec.ts
@@ -1,9 +1,8 @@
 import { Component, viewChild, viewChildren } from '@angular/core';
 import { type ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
-
-import { SbbFormField } from '../form-field/form-field';
-import { SbbOption } from '../option/option';
+import { SbbFormField } from '@sbb-esta/lyne-angular/form-field/form-field';
+import { SbbOption } from '@sbb-esta/lyne-angular/option/option';
 
 import { SbbAutocomplete } from './autocomplete';
 import { SbbAutocompleteTrigger } from './autocomplete-trigger';
