+++ src/angular/breadcrumb/breadcrumb-group/breadcrumb-group.spec.ts
@@ -1,9 +1,8 @@
 import { Component, ElementRef, viewChildren } from '@angular/core';
 import { type ComponentFixture, TestBed } from '@angular/core/testing';
-import { SbbBreadcrumb } from '@sbb-esta/lyne-angular/breadcrumb/breadcrumb';
 import type { SbbBreadcrumbElement } from '@sbb-esta/lyne-elements/breadcrumb/breadcrumb.js';
 
-import { SbbBreadcrumbGroup } from './breadcrumb-group';
+import { SbbBreadcrumbGroup, SbbBreadcrumb } from '..';
 
 describe('sbb-breadcrumb-group', () => {
   let fixture: ComponentFixture<TestComponent>, component: TestComponent;
