+++ src/angular/chip/chip-group/chip-group.spec.ts
@@ -2,12 +2,11 @@ import type { ElementRef, Signal } from '@angular/core';
 import { Component, viewChild } from '@angular/core';
 import { type ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
+import { SbbAutocomplete } from '@sbb-esta/lyne-angular/autocomplete/autocomplete';
 import { SbbFormField } from '@sbb-esta/lyne-angular/form-field/form-field';
+import { SbbOption } from '@sbb-esta/lyne-angular/option/option';
 import type { SbbChipInputTokenEndEventDetails } from '@sbb-esta/lyne-elements/chip/chip-group.js';
 import type { SbbChipElement } from '@sbb-esta/lyne-elements/chip/chip.js';
-
-import { SbbAutocomplete } from '../../autocomplete';
-import { SbbOption } from '../../option/option';
 
 import { SbbChipGroup } from './chip-group';
