--- src/angular/table/table-wrapper/index.ts
@@ -1,1 +1,0 @@
-export * from './table-wrapper';
