--- src/angular/table/table-wrapper/ng-package.json
@@ -1,5 +1,0 @@
-{
-  "lib": {
-    "entryFile": "index.ts"
-  }
-}
