+++ src/showcase/stories/angular/table/table-wrapper/table-wrapper.stories.ts
@@ -2,8 +2,7 @@ import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { SbbFormField } from '@sbb-esta/lyne-angular/form-field/form-field';
 import { SbbPaginator } from '@sbb-esta/lyne-angular/paginator/paginator';
-import { SbbTableDataSource, SbbTableModule } from '@sbb-esta/lyne-angular/table';
-import { SbbTableWrapper } from '@sbb-esta/lyne-angular/table/table-wrapper';
+import { SbbTableDataSource, SbbTableModule, SbbTableWrapper } from '@sbb-esta/lyne-angular/table';
 import type { Args, Meta } from '@storybook/angular';
 import { applicationConfig, moduleMetadata } from '@storybook/angular';
 import type { ArgTypes, InputType, StoryContext } from 'storybook/internal/types';
